requirements.txt
Click
google-auth-oauthlib
google-cloud-bigquery
jinja
pyperclip
pyyaml
readchar
rich